Added support for target_feature atomics on wasm32
The target_feature atomics enabled multithread on the wasm32.

The devices in wgpu are not Send or Sync when the target_feature
atmoics is enabled as they cannot be shared across threads.

This PR enables cubecl to run wgpu on a dedicated thread and then
communicate to this thread using channels.